Separated into 3 files, restructured code
There are now two class files and the main file. All the code is now in
a class-centric format with each class being logically separate.
