+++ startMenu.py
@@ -1,0 +1,192 @@
+import pygame
+from pygame.locals import *
+import os
+import sys
+
+class Start(object):
+    def __init__(self, screen, font_op):
+        self.screen = screen
+        self.font_op = font_op
+        self.gameActive = 0  #Is the game running or not
+        #self.quitGame = False
+        self.soundRunning = 0
+        self.load_images()
+        self.load_sounds()
+        self.menuOn = True
+        self.rulesOn = False
+        self.clock = pygame.time.Clock()
+
+    def load_images(self):
+        self.img_logo = pygame.image.load(os.path.join("img","logo.png")).convert_alpha()
+
+        self.img_menu_bg = pygame.image.load(os.path.join("img","menu_bg4.png")).convert()
+                 
+        #self.img_arrow = pygame.image.load(os.path.join("img","arrow.png")).convert_alpha()
+        #self.img_arrow_left = pygame.image.load(os.path.join("img","arrowLeft.png")).convert_alpha()
+        
+        self.img_icon_start = pygame.image.load(os.path.join("img","Dice2b.png")).convert_alpha()
+        self.img_icon_resume = pygame.image.load(os.path.join("img","thumb_up2.png")).convert_alpha()
+        self.img_icon_rules = pygame.image.load(os.path.join("img","book-brownb.png")).convert_alpha()
+        self.img_icon_options = pygame.image.load(os.path.join("img","Gear_01b.png")).convert_alpha()
+        self.img_icon_exit = pygame.image.load(os.path.join("img","Exitb.png")).convert_alpha() #Game exit
+        #self.img_icon_exit_x = pygame.image.load(os.path.join("img","exit.png")).convert_alpha() #Rules exit
+
+    def load_sounds(self): 
+        self.sound_intro = pygame.mixer.Sound(os.path.join('sound','radio.wav'))
+        self.sound_start_menu = pygame.mixer.Sound(os.path.join('sound','start_menu.wav'))
+
+    def splash(self):
+        start_time = pygame.time.get_ticks()/1000.0
+        #Remove multiple line comment below to view splash screen
+        text_color = [230,230,230]
+        
+        while True:
+            self.clock.tick(30)
+            self.screen.fill((0,0,0))
+            time = pygame.time.get_ticks()/1000. - start_time
+            
+            if time > 1 and time < 6:            
+                for i in range(3):
+                    if text_color[i] < 255:
+                        text_color[i] -= 1.3
+                text_title = self.font_op(72,"berlin").render("Mastering MSU",True,text_color)
+                #text_team1 = font_op(24,"helvetica").render("a Team 1 game",True,text_color)
+             
+                self.screen.blit(self.img_logo,(self.screen.get_width()/2-0.5*self.img_logo.get_width(),220-0.5*self.img_logo.get_height()))
+                self.screen.blit(text_title,(self.screen.get_width()/2-0.5*text_title.get_width(),280))
+                #self.screen.blit(text_team1,(self.screen.get_width()/2-0.5*text_team1.get_width(),380))        
+            if time > 1 and self.soundRunning == 0:
+                self.sound_intro.play()
+                self.sound_intro.fadeout(6000)
+                self.soundRunning = 1
+            if time > 6.5:
+                self.soundRunning = 0
+                break
+            
+            pygame.display.update()
+
+    def leaveGame(self):
+        pygame.quit()
+        sys.exit()
+
+    def highlight(self):
+        mouseX,mouseY = pygame.mouse.get_pos() 
+        if mouseX > self.screen.get_width()/2-107\
+           and mouseX < self.screen.get_width()/2-57+self.text_start.get_width()+5\
+           and mouseY > 195 and mouseY < 200+self.text_start.get_height():
+            self.text_start = self.font_op(32,"berlin").render("Start Game",True,(255,255,255))
+        elif self.gameActive == 1:
+            self.text_start = self.font_op(32,"berlin").render("Start Game",True,(220,146,40))  
+            if mouseX > self.screen.get_width()/2-107\
+               and mouseX < self.screen.get_width()/2-57+self.text_resume.get_width()+5\
+               and mouseY > 270 and mouseY < 275+self.text_resume.get_height():
+                self.text_resume = self.font_op(32,"berlin").render("Resume Game",True,(255,255,255))
+            else:
+                self.text_resume = self.font_op(32,"berlin").render("Resume Game",True,(220,146,40))
+        else:
+            self.text_start = self.font_op(32,"berlin").render("Start Game",True,(220,146,40))    
+        if mouseX > self.screen.get_width()/2-107\
+           and mouseX < self.screen.get_width()/2-57+self.text_rules.get_width()+5\
+           and mouseY > 270+75*self.gameActive\
+           and mouseY < 275+self.text_rules.get_height()+75*self.gameActive:
+            self.text_rules = self.font_op(32,"berlin").render("Rules",True,(255,255,255))
+        else:
+            self.text_rules = self.font_op(32,"berlin").render("Rules",True,(220,146,40))
+        if mouseX > self.screen.get_width()/2-107\
+           and mouseX < self.screen.get_width()/2-57+self.text_options.get_width()+5\
+           and mouseY > 345+75*self.gameActive\
+           and mouseY < 350+self.text_options.get_height()+75*self.gameActive:
+            self.text_options = self.font_op(32,"berlin").render("Options",True,(255,255,255))
+        else:
+            self.text_options = self.font_op(32,"berlin").render("Options",True,(220,146,40))       
+        if mouseX > self.screen.get_width()/2-107\
+           and mouseX < self.screen.get_width()/2-57+self.text_exit.get_width()+5\
+           and mouseY > 420+75*self.gameActive\
+           and mouseY < 425+self.text_exit.get_height()+75*self.gameActive:
+            self.text_exit = self.font_op(32,"berlin").render("Exit",True,(255,255,255))
+        else:
+            self.text_exit = self.font_op(32,"berlin").render("Exit",True,(220,146,40))
+
+    def buttonClick(self):
+        mouseX,mouseY = pygame.mouse.get_pos()
+        if mouseX > self.screen.get_width()/2-107\
+           and mouseX < self.screen.get_width()/2-57+self.text_rules.get_width()+5\
+           and mouseY > 270+75*self.gameActive\
+           and mouseY < 275+self.text_rules.get_height()+75*self.gameActive:
+            self.menuOn = False
+            self.rulesOn = True
+        elif mouseX > self.screen.get_width()/2-107\
+             and mouseX < self.screen.get_width()/2-57+self.text_exit.get_width()+5\
+             and mouseY > 420+75*self.gameActive\
+             and mouseY < 425+self.text_exit.get_height()+75*self.gameActive:
+            self.leaveGame()
+
+    def imageDisplay(self):
+        self.screen.blit(self.img_menu_bg, (0, 0))
+        self.screen.blit(self.text_header_bg,(self.screen.get_width()/2-3-0.5*self.text_header_bg.get_width()+4,20))
+        self.screen.blit(self.text_header,(self.screen.get_width()/2-3-0.5*self.text_header.get_width(),16))
+
+        self.screen.blit(self.text_start_bg,(self.screen.get_width()/2-55,202))
+        self.screen.blit(self.text_start,(self.screen.get_width()/2-57,200))
+        self.screen.blit(self.text_rules_bg,(self.screen.get_width()/2-55,277+75*self.gameActive))
+        self.screen.blit(self.text_rules,(self.screen.get_width()/2-57,275+75*self.gameActive))
+        self.screen.blit(self.text_options_bg,(self.screen.get_width()/2-55,352+75*self.gameActive))
+        self.screen.blit(self.text_options,(self.screen.get_width()/2-57,350+75*self.gameActive))
+        self.screen.blit(self.text_exit_bg,(self.screen.get_width()/2-55,427+75*self.gameActive))
+        self.screen.blit(self.text_exit,(self.screen.get_width()/2-57,425+75*self.gameActive))
+
+        self.screen.blit(self.img_icon_start,(self.screen.get_width()/2-112,195))
+        self.screen.blit(self.img_icon_rules,(self.screen.get_width()/2-112,270+75*self.gameActive))
+        self.screen.blit(self.img_icon_options,(self.screen.get_width()/2-112,345+75*self.gameActive))
+        self.screen.blit(self.img_icon_exit,(self.screen.get_width()/2-112,420+75*self.gameActive))
+
+    def menu(self):
+        self.text_header = self.font_op(90,"berlin").render("Mastering MSU",True,(255,255,255))
+        self.text_header_bg = self.font_op(90,"berlin").render("Mastering MSU",True,(0,0,0))
+        self.text_start = self.font_op(32,"berlin").render("Start Game",True,(255,255,255))
+        self.text_start_bg = self.font_op(32,"berlin").render("Start Game",True,(0,0,0))
+        self.text_resume = self.font_op(32,"berlin").render("Resume Game",True,(255,255,255))
+        self.text_resume_bg = self.font_op(32,"berlin").render("Resume Game",True,(0,0,0))
+        self.text_rules = self.font_op(32,"berlin").render("Rules",True,(255,255,255))
+        self.text_rules_bg = self.font_op(32,"berlin").render("Rules",True,(0,0,0))
+        self.text_options = self.font_op(32,"berlin").render("Options",True,(255,255,255))
+        self.text_options_bg = self.font_op(32,"berlin").render("Options",True,(0,0,0))
+        self.text_exit = self.font_op(32,"berlin").render("Exit",True,(255,255,255))
+        self.text_exit_bg = self.font_op(32,"berlin").render("Exit",True,(0,0,0))
+
+        #menu = True
+        #rulesActive = False
+
+        while self.menuOn:
+            self.clock.tick(30)
+            self.screen.fill((240,240,240))
+
+            if self.soundRunning == 0:
+                self.sound_start_menu.play(-1) 
+                self.soundRunning = 1
+
+            for event in pygame.event.get():
+                if event.type == QUIT:
+                    self.leaveGame()
+                #if event.type == KEYDOWN:
+                    #if event.key == K_ESCAPE:
+                        #if rulesActive == True:
+                            #rulesActive = False
+                        #if self.gameActive == 1:
+                            #menu = False
+                            
+                if event.type == MOUSEMOTION:  #Highlight text on hover
+                    self.highlight()
+                        
+                if event.type == MOUSEBUTTONDOWN:  #Perform action on click
+                    self.buttonClick()
+            self.imageDisplay()
+
+            if self.gameActive == 1:
+                self.screen.blit(self.text_resume_bg,(self.screen.get_width()/2-55,277))
+                self.screen.blit(self.text_resume,(self.screen.get_width()/2-57,275))
+                self.screen.blit(self.img_icon_resume,(self.screen.get_width()/2-112,270))
+            
+            pygame.display.update()
+        if self.rulesOn:
+            return "rules"
